# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KepRoop API Documentation",
    "version": "1.0"
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    },
    {
      "CookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Prefix with 'Api-Key ' or 'Bearer kp_'"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Browser Session Cookie (Automatically used if logged in)"
      }
    }
  },
  "paths": {
...
    "/api/albums/{id}": {
      "get": {
        "tags": [
          "Albums"
        ],
        "summary": "Get album details",
        "description": "Get album details including images and folders.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Album details"
          },
          "404": {
            "description": "Album not found"
          }
        }
      },
      "put": {
        "tags": [
          "Albums"
        ],
        "summary": "Update album",
        "description": "Update album metadata.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  },
                  "albumDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Album updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Albums"
        ],
        "summary": "Delete album",
        "description": "Delete an album and all its contents.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Album deleted"
          }
        }
      }
    },
    "/api/albums/{id}/trash": {
      "get": {
        "tags": [
          "Albums"
        ],
        "summary": "View trash",
        "description": "List deleted images in the album.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of deleted images"
          }
        }
      },
      "delete": {
        "tags": [
          "Albums"
        ],
        "summary": "Empty trash / Permanently delete",
        "description": "Permanently delete specific images from trash.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "imageIds"
                ],
                "properties": {
                  "imageIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Images permanently deleted"
          }
        }
      }
    },
    "/api/albums": {
      "get": {
        "tags": [
          "Albums"
        ],
        "summary": "List albums",
        "description": "List albums with filtering, sorting, and pagination.",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            },
            "description": "Cursor for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 12
            },
            "description": "Number of items to return"
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "mine",
                "shared",
                "favorites"
              ],
              "default": "all"
            },
            "description": "Filter albums by ownership"
          },
          {
            "in": "query",
            "name": "visibility",
            "schema": {
              "type": "string",
              "enum": [
                "public",
                "private"
              ]
            },
            "description": "Filter by visibility"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by title or description"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "albumDate"
            },
            "description": "Sort field"
          },
          {
            "in": "query",
            "name": "sortDir",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort direction"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of albums"
          }
        }
      },
      "post": {
        "tags": [
          "Albums"
        ],
        "summary": "Create album",
        "description": "Create a new album.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": [
                      "public",
                      "private"
                    ]
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "coverImageKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Album created successfully"
          }
        }
      }
    },
...
  },
...
}